Moves over non-runtime required tests from ImportHelpersTest (Runtime) to ImportHelpersTest (Editor)
